+++ pkg/controllers/kusciatask/handler/finished_handler.go
@@ -27,7 +27,6 @@ import (
 	corelisters "k8s.io/client-go/listers/core/v1"
 
 	"github.com/secretflow/kuscia/pkg/common"
-	"github.com/secretflow/kuscia/pkg/controllers/kusciatask/dependencies"
 	kusciaapisv1alpha1 "github.com/secretflow/kuscia/pkg/crd/apis/kuscia/v1alpha1"
 	kusciaclientset "github.com/secretflow/kuscia/pkg/crd/clientset/versioned"
 
@@ -43,7 +42,7 @@ type FinishedHandler struct {
 }
 
 // NewFinishedHandler returns a FinishedHandler instance.
-func NewFinishedHandler(deps *dependencies.Dependencies) *FinishedHandler {
+func NewFinishedHandler(deps *Dependencies) *FinishedHandler {
 	return &FinishedHandler{
 		kubeClient:      deps.KubeClient,
 		kusciaClient:    deps.KusciaClient,
