+++ pkg/controllers/kusciatask/handler/running_handler_test.go
@@ -28,7 +28,6 @@ import (
 	kubefake "k8s.io/client-go/kubernetes/fake"
 
 	"github.com/secretflow/kuscia/pkg/common"
-	"github.com/secretflow/kuscia/pkg/controllers/kusciatask/dependencies"
 	kusciaapisv1alpha1 "github.com/secretflow/kuscia/pkg/crd/apis/kuscia/v1alpha1"
 	kusciafake "github.com/secretflow/kuscia/pkg/crd/clientset/versioned/fake"
 	kusciainformers "github.com/secretflow/kuscia/pkg/crd/informers/externalversions"
@@ -182,7 +181,7 @@ func TestRunningHandler_Handle(t *testing.T) {
 			trgInformer.Informer().GetStore().Add(tt.trg)
 
 			go kubeInformersFactory.Start(wait.NeverStop)
-			deps := dependencies.Dependencies{
+			deps := Dependencies{
 				KubeClient:       kubeClient,
 				KusciaClient:     kusciaClient,
 				TrgLister:        trgInformer.Lister(),
