+++ pkg/controllers/kusciatask/handler/succeeded_handler_test.go
@@ -23,7 +23,6 @@ import (
 	kubeinformers "k8s.io/client-go/informers"
 	kubefake "k8s.io/client-go/kubernetes/fake"
 
-	"github.com/secretflow/kuscia/pkg/controllers/kusciatask/dependencies"
 	kusciaapisv1alpha1 "github.com/secretflow/kuscia/pkg/crd/apis/kuscia/v1alpha1"
 	kusciafake "github.com/secretflow/kuscia/pkg/crd/clientset/versioned/fake"
 	kusciainformers "github.com/secretflow/kuscia/pkg/crd/informers/externalversions"
@@ -43,7 +42,7 @@ func TestSucceededHandler_Handle(t *testing.T) {
 	kusciaInformerFactory := kusciainformers.NewSharedInformerFactory(kusciaClient, 0)
 	go kubeInformersFactory.Start(wait.NeverStop)
 
-	deps := dependencies.Dependencies{
+	deps := Dependencies{
 		KubeClient:      kubeClient,
 		KusciaClient:    kusciaClient,
 		TrgLister:       kusciaInformerFactory.Kuscia().V1alpha1().TaskResourceGroups().Lister(),
